refactor: extract shared test fakes and helpers into tests/helpers.py
Duplicate fake objects (FakeLLM, FakePriceProvider, etc.) and factory
functions (make_llm_registry, make_data_registry, etc.) were defined
independently across 6 test files. Extract them into a shared
tests/helpers.py module to eliminate duplication and provide a single
source of truth for test infrastructure.

Changes:
- Create tests/helpers.py with shared fakes, factories, and helpers
- Update test_agents.py, test_engine.py, test_intent.py,
  test_trade_thesis.py, test_context.py to import from helpers
- Add "tests" to pythonpath in pyproject.toml for cross-module imports
- ~120 lines of duplicate code removed across test files
