+++ tests/conftest.py
@@ -1,0 +1,1 @@
+"""Root conftest — makes the tests/ directory importable for shared helpers."""
